+++ domain/model_secret.go
@@ -8,6 +8,7 @@ type Secret struct {
 	Namespace string            `json:"namespace"`
 	Checksum  string            `json:"checksum"`
 	Data      map[string]string `json:"data"`
+	Labels    map[string]string `json:"labels"`
 }
 
 // Repository interface
@@ -16,9 +17,9 @@ type Repository interface {
 	// CheckSecretK8S return a annotation to validate the secret already created
 	CheckSecretK8S(name string, namespace string) (string, string, error)
 	// CreateSecretK8S creates a new secret
-	CreateSecretK8S(name string, checksum string, namespace string, data map[string]string) (string, error)
+	CreateSecretK8S(name string, checksum string, namespace string, data, labels map[string]string) (string, error)
 	// UpdateSecretK8S updates an already created secret
-	UpdateSecretK8S(name string, checksum string, namespace string, data map[string]string) (string, error)
+	UpdateSecretK8S(name string, checksum string, namespace string, data, labels map[string]string) (string, error)
 	// DeleteSecretK8S deletes a secret
 	DeleteSecretK8S(name string, namespace string) (string, error)
 }